WAD-master/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.10.11=h06a4308_0
certifi=2022.9.24=py37h06a4308_0
cffi=1.15.1=py37h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=38.0.1=py37h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
fontconfig=2.13.1=hef1e5e3_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.4=py37h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.1=py37h06a4308_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=py37h295c915_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.4.0=hecacb30_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.3=py37h06a4308_0
matplotlib-base=3.5.3=py37hf590b9c_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pillow=9.2.0=py37hace64e9_1
pip=22.2.2=py37h06a4308_0
ply=3.11=py37_0
protobuf=3.20.1=py37h295c915_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.15.7=py37h6a678d5_1
pyqt5-sip=12.11.0=py37h6a678d5_1
pysocks=1.7.1=py37_1
python=3.7.13=haa1d7c7_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.12.0=py3.7_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
requests=2.28.1=py37h06a4308_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37h6c91a56_2
setuptools=65.5.0=py37h06a4308_0
sip=6.6.2=py37h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h5082296_0
tensorboardx=2.2=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.12.0=py37_cu102
torchvision=0.13.0=py37_cu102
tornado=6.2=py37h5eee18b_0
tqdm=4.64.1=py37h06a4308_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.12=py37h06a4308_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
